fix: change capacity to long

+++ sdk/compute/mgmt/src/main/java/com/azure/management/compute/implementation/VirtualMachineScaleSetImpl.java
@@ -1033,8 +1033,8 @@ public class VirtualMachineScaleSetImpl
     }
 
     @Override
-    public VirtualMachineScaleSetImpl withCapacity(int capacity) {
-        this.inner().sku().withCapacity((long) capacity);
+    public VirtualMachineScaleSetImpl withCapacity(long capacity) {
+        this.inner().sku().withCapacity(capacity);
         return this;
     }
